Fix more stringent tests that failed for python3
The valuelist.render refractor also tightened up several tests which
test if an exception was raised. The changes added a test of the message
returned by the exception. As the evaluation order of independant reclass
parameters is not defined it's possible to get one of several different
error message for some tests. python2 uses the same evaluation order on
each reclass run but for python3 the order can and does change between
reclass runs over the same data.

The tests are fixed by allowing any of the valid error messages to pass
the test.