Allow for class mappings to wildcard match against either the node name or node file path
By default a class mapping will match against the name of a node:

class_mappings:
  - test*        alpha

This will match all nodes names starting with test.

A possibly unintended side effect of the original implementation allowed
for matches against the path of the node file. So for a node file
'test/node1.yml' a match:

class_mappings:
  - test/*        alpha

would work. A reworking of the class mappings implementation ment this
matching stopped working. The type of matching done by the class mapping
is now controlled by the configuration option class_mappings_match_path.
If false (the default) the match is done against the node name and if true
the match is done again the path.