Avoid modifying context.parameters during class resolution
The initial implementation to allow parameter interpolation to be used
in class names ([#2]) resulted in `uri` fields peppered all over the
resulting parameters object, due to the [try-catch] when setting
`cur.uri` introduced in `Parameters._merge_dict`, which would put a key
`uri` into the merged parameters dict, if the passed in `cur` is a plain
dict instead of a `ParameterDict`.

As far as I can figure out, this happens when multiple classes are
defined using parameters. I suspect that in this case, the
`context.parameters` object is initialised for interpolation multiple
times which leads to odd behaviour and results in a call to
`Parameters._merge_dict` where cur is a plain dict instead of a
`ParameterDict`.

By creating a deepcopy of `context.parameters` when interpolating the
context parameters during class resolution, we completely avoid the
situation that we try to (re)initialise interpolation on
`context.parameters`, and the original copy of `context.parameters`
remains untouched during class resolution.

With this change, we can remove the try-catch when setting `cur.uri` in
`Parameters._merge_dict`, as we now always have properly only-once
initialised Parameters objects for interpolation.

[#2]: https://github.com/kapicorp/reclass/pull/2
[try-catch]: https://github.com/kapicorp/reclass/blob/2585bfc620ddd8c2cd7d159c4eb6d117a526a819/reclass/datatypes/parameters.py#L188-L191