Add test which illustrates the potential bug when merging applications
According to [the reclass documentation], prefixing an application with
"~" will drop it from the list of applications, if it's present.

This commit adds a new test which uses this behavior to replace "app4"
with "app5" for environment A. The way classes are included in the test
is a bit contrived, but tries to present a simplified version of the
class includes we use in [Commodore](https://github.com/projectsyn/commodore).

The current reclass implementation fails to remove "app4" for node
"A_node", even though the "environments.envA" class is included after
the "applications.shared" class.

[the reclass documentation]: https://reclass.pantsfullofunix.net/operations.html#yaml-fs-storage