Adjust _recurse_entity to process removal in applications depth-first
With the previous implementation, the `applications` array was handled
individually for each ancestor of an entity. This appears to contradict
the description of how [data merging] is performed in the reclass
documentation. The documentation specifies that

> Next, reclass recursively descends each class, looking at the classes
> it defines, and so on, until a leaf node is reached, i.e. a class that
> references no other classes.

> Now, the merging starts. At every step, the list of applications and
> the set of parameters at each level is merged into what has been
> accumulated so far.

So, classes are traversed depth-first in the order they're given in the
`classes` array.

The implementation correctly merges parameters and addititons to the
applications array depth-first in the order the classes are given.

However, the current implementation doesn't process removal for the
`applications` array in a truly depth-first manner. The cause is that
removal of entries in the `applications` array is processed immediately
when a "~"-prefixed entry is found. Therefore removal of entries is only
possible at different levels of the hierarchy, as the current
implementation starts the traversal of each ancestor of a class with the
same input data, and merges the result of the traversal into the
accumulated result after the traversal completes.

As noted above, this post-traversal merging works fine for purely
additive operations, but doesn't correctly handle subtractive
operations, as those are processed immediately when they're encountered,
and are not visible at the time the merging actually takes place.

This commit changes `_recurse_entity` to instead start each traversal
using the already accumulated data as the base for the traversal. With
this change, if a "~"-prefixed entry is found in an ancestor, the
recursion has a fully-updated base on which it can process the removal.
This allows a class to remove entries in `applications` which were added
by classes that were listed before it in the `classes` array.

By changing how we update `merge_base` during recursive calls to
`_recurse_entity`, we also need to refactor the code implemented to
support references in class names. With the new approach to updating
`merge_base`, we can create a deep copy of `merge_base.parameters` which
we can use to resolve and references using the already accumulated set
of parameters.

[data merging]: https://reclass.pantsfullofunix.net/operations.html#data-merging